Mask unwanted bits in the result of ~.
GCC optimizes away ~*gpio used as a condition, possibly because
apparently the ~ operator ignores the operand type unlike other
operators.  It seems the operand is always cast to signed in and the
result is a signed int.  This was causing a very tricky bug to occur
that could only be seen in the resulting assembly.  If *gpio was
marked volatile, it would still only compile to "ld <reg>, gpio" with
optimisation.

+++ Sensorino/Sensorino.cpp
@@ -255,7 +255,7 @@ static void SensorinoGPIOISR(int port, uint8_t new_val) {
 
     for (volatile uint8_t *gpio = &pcint_to_gpio[port << 3]; diff;
             diff >>= 1, gpio++)
-        if ((diff & 1) && ~*gpio) {
+        if ((diff & 1) && (uint8_t) ~*gpio) {
             if (gpio_obj_mask & (1 << *gpio)) {
                 GenIntrCallback *cb = (GenIntrCallback *) gpio_handler[*gpio];
                 cb->call(*gpio);
